Pre-handlers y otros - Separar message_cmd y algunas funciones - !lockbot ahora es un prehandler - SW-handlers pueden usar $PX - Quitar verificación de lock de command_handler - Poner otros handlers dentro de try
